add delete prompts

+++ app/admin/resources/detail/controller.js
@@ -4,8 +4,10 @@ export default Ember.Controller.extend({
   actions: {
     deleteOrg() {
       const no = this.model;
-      no.destroyRecord();
-      this.transitionToRoute('admin');
+      if (confirm('Are you for real?')) {
+        no.destroyRecord();
+        this.transitionToRoute('admin.resources');
+      }
     }
 
   }
